Initial commit, version 0.3.0
This version can listen for posts with a custom delay and translate,
with a limited capacity, words into numbers to filter out posts below a
certain price.

At the moment, it technically can't differentiate between buy/sell. In
the future, it could be modified to include filtering for bells below a
certain amount for buy filtering.